+++ docs/CHANGELOG.md
@@ -2,6 +2,18 @@
 > windowsboy111
 
 ## [Unreleased]
+### Added
+- now using setuptools in `setup.py` and added install requires
+- when restart / shutdown it will auto-save changes
+- now by default it will not say that command not found
+### Optimized
+- compressed bot.py
+### Removed
+- spamming detection system
+- member welcoming and bot setup, cuz they are annoying tbh with you, and they are optional for the bot to work
+### Fixed
+- the bot mention prefix is now dynamic
+- auto create lastmsg file when it does not exist
 
 ## [2.1.4]
 ### Added
@@ -18,7 +30,6 @@
 - #22: `/poll end all` will not crash when there are no unended polls
 - #23
 - mention of Merlin was not prefix
-
 
 ## [2.1.3]
 ### Added